Reorganize constants, add headers, and clean up files
