+++ genai-perf/genai_perf/inputs/input_constants.py
@@ -54,29 +54,48 @@ class OutputFormat(Enum):
         return self.name.lower()
 
 
-OPEN_ORCA = "openorca"
+###########################
+# General Parameters
+###########################
+DEFAULT_INPUT_DATA_JSON = "inputs.json"
+DEFAULT_RANDOM_SEED = 0
+
+
+####################
+# URL Constants
+####################
 CNN_DAILY_MAIL = "cnn_dailymail"
+OPEN_ORCA = "openorca"
 OPEN_ORCA_URL = "https://datasets-server.huggingface.co/rows?dataset=Open-Orca%2FOpenOrca&config=default&split=train"
 CNN_DAILYMAIL_URL = "https://datasets-server.huggingface.co/rows?dataset=cnn_dailymail&config=1.0.0&split=train"
-DEFAULT_INPUT_DATA_JSON = "inputs.json"
+dataset_url_map = {OPEN_ORCA: OPEN_ORCA_URL, CNN_DAILY_MAIL: CNN_DAILYMAIL_URL}
+
+###########################
+# Default Prompt Parameters
+###########################
 DEFAULT_STARTING_INDEX = 0
 MINIMUM_STARTING_INDEX = 0
 DEFAULT_LENGTH = 100
 MINIMUM_LENGTH = 1
 DEFAULT_TENSORRTLLM_MAX_TOKENS = 256
 DEFAULT_BATCH_SIZE = 1
-DEFAULT_RANDOM_SEED = 0
 DEFAULT_PROMPT_TOKENS_MEAN = 550
 DEFAULT_PROMPT_TOKENS_STDDEV = 0
 DEFAULT_OUTPUT_TOKENS_MEAN = -1
 DEFAULT_OUTPUT_TOKENS_STDDEV = 0
 DEFAULT_NUM_PROMPTS = 100
+
+###########################
+# Default Image Parameters
+###########################
 DEFAULT_IMAGE_WIDTH_MEAN = 100
 DEFAULT_IMAGE_WIDTH_STDDEV = 0
 DEFAULT_IMAGE_HEIGHT_MEAN = 100
 DEFAULT_IMAGE_HEIGHT_STDDEV = 0
+
+###########################
+# Default JSON Parameters
+###########################
 EMPTY_JSON_IN_VLLM_PA_FORMAT: Dict = {"data": []}
 EMPTY_JSON_IN_TENSORRTLLM_PA_FORMAT: Dict = {"data": []}
 EMPTY_JSON_IN_OPENAI_PA_FORMAT: Dict = {"data": []}
-
-dataset_url_map = {OPEN_ORCA: OPEN_ORCA_URL, CNN_DAILY_MAIL: CNN_DAILYMAIL_URL}
